Fix underlinking of hilti-rt library.
Previously, functions exported by justrx where not linked into hilti-rt
leading to unresolved linktime dependencies when linking code requiring
some forms of regular expressions. This was due to a quirk in how CMake
handles transitive dependencies of OBJECT libraries.

In this case we had hilti-rt depend on the OBJECT library
hilti-rt-objects which in turn depends on the OBJECT library
jrx-objects. Since only usage requirements, but not object files of
OBJECT libraries transitive, this caused the symbols from jrx-objects to
not be included into hilti-rt, see
https://cmake.org/cmake/help/latest/command/target_link_libraries.html#linking-object-libraries.
Adding a direct link between hilti-rt and jrx-objects resolves this.