Fix initialization of &default struct parameters.
In some cases, the corrrespondsing struct members would not be
initialized.

In addition, this commit also revamps generation of the struct's
custom constructors: We move that from the HILTI codegen over to the
CXX layer, which makes it quite a bit cleaner. That way, we also don't
need a separate internal __init() method anymore, but can just have
the various constructors do what they need to do.

Closes #163.