Activate runtime location tracking.
This enables us to include location information when throwing runtime
exceptions. We already had the code in place to do the tracking, but
it required explicit activation through a debug option. We know
generally insert the calls to record the current location into the
generated code, seems well worth it. (We could opt to skip it when
optimizing, not sure yet).

It also moves the storage for the current location from being a global
variable into the execution context; that makes it thread safe.

Closes #119.