Add generated docs to the tree.
This patch adds generated documentation files to the main tree so that
one can build the full documentation without needing to build Spicy
itself. This is needed so that e.g., readthedocs can generate docs.

Ideally we would invoke the script added here as part of a pre-commit
hook and validate that no files were modified. This is currently not
possible as `doc/scripts/spicy-doc-to-rst` does not behave
deterministically (running it is e.g., not idempotent).