Do not reuse names with different types.
The way name lookup works in Python we were relying on the ability to
change the dynamic type of variables in that we were binding different
types to variables of the same name. The code was still behaving as
expected, but this pattern is confusing (e.g., to human readers or
static type checkers).