Require spicyfing exact command to run in Docker containers.
We use this to decouple building of the base image which shared between
a user-accessible container and also built and stored in a registry in
Cirrus CI. The container containing the user-accessible Spicy Docker
image can then depend on the shared base image.

In order to parameterize the Spicy Docker image build process outside of
the Dockerfile we now require a third argument to
`docker/docker-helper` which should contain the command to run in the
base image.