Remove unused `SPICY_BUILD_DIRECTORY` env var in doc setup.
In `074f6fc69f7` we moved all generated doc files into the tree to make
building the documentation independent from building the code. We left
behind an unused environment variable `SPICY_BUILD_DIRECTORY` in the
doc's Sphinx config which was previously used to communicate the
location of required binaries, but which is now unused.

This patch removes use of the environment variable to enable doc
builders like readthedocs to build to render te documentation without
special setup.