Ignore package links in documentation linkcheck.
We check links during the `build` phase, but generate package artifacts
in a follow-up step. If for some reason package artifacts are absent in
Cirrus CI's artifact cache, this can lead to situations where the link
check fails and we never repopulate the artifacts.

For now just ignore links to the artifacts when performing the
documentation link check.