Fix `hilti::rt::to_string` for character literals.
We previously did not provide a `to_string` specialization for character
literals and instead resolved to an integer specialization. For that the
literal had to decay to a `char` pointer which could then convert to an
integer. This lead to us rendering not the expected result.

This patch adds a `to_string` specialization for character literals so
that they are rendered just like the equivalent `std::string`.