Reject container constructors with inconsistent types, and coerce elements once the type is known.
For a list constructor [x_1,...,x_n], we used to take the type of the
first element as the list's element type. Now we check all elements if
their types match; only if so, we use that type. Otherwise we treat
the element type as unknown, which may then lead to validation errors.

Once we know the element type of a list constructor (either directly,
or inferred per above), we now also coerce all the elements
accordingly.

We further unify this logic across all container constructors
(list/map/set/vector).

Closes #196.