Fix semantics of safe mutating access to `hilti::rt::Vector` elements.
Semantically we want to allow getting mutable references to `Vector`
elements which are not (yet) present, e.g., to bind references to them
before assigning later in the control flow. In such cases any range
checks should be performed lazily when actually mutating the referenced
value (as opposed to: when obtianing the reference). This is different
from cases where we get references which are fundamentally not mutable
(i.e., which are strictly speaking rvalues); in such cases the range
check has to be performed when the reference is obtained.

This patch implements this behavior by making `Vector`s subscript
operator aware on whether it produces an lvalue or rvalue. We also clean
up an existing TODO in `Vector`'s `AssignProxy`.