Fix AST inconsistency with how modules were stored in scopes.
Normally all scope entries should be pointing to declarations, but we
made an exception for "Module", which was inserted into the scope
directly. This commit introduces a separate "declaration::Module"
wrapper so that we can insert that into the scope, avoiding that
exception.

With that in place, we can then easily catch if an ID refers directly
to a module, which has no use case.

For the record: I did consider removing the current "Module" node
class altogether, and simply replacing it with "declaration::Module".
However, "Module" is special in that it's our only mutable AST node,
and it didn't feel right to mix that in with the other, non-mutable
declarations.

Closes #248.