Avoid leaking passwords into command lines in `docker/Makefile` [skip ci].
This patch changes `docker/Makefile` so that the Docker Hub password is
not passed as on the command line anymore which could be inspected by
anyone with access to process table of the machine the job is running.
Instead we now read the password from a file and pass it via a pipe into
the login command.