Add support for "&until_including" when parsing bytes.
The error message one got when trying to use &until_including with a
field of type "bytes" would claim that "bytes" was indeed supported.
It was lying. To fix the error message, this commit adds support for
that feature:

    # cat test.spicy

    module Test;

    public type X = unit {
        x: bytes &until_including=b"567";
        on %done { print self; }
    };

    # printf '1234567890' | spicy-driver test.spicy
    [$x=b"1234567"]

The commits also cleans up validation for attributes on bytes fields.

Closes #269.