Fix Bison parser's attribute association for regular expressions.
In regex fields, the Bison parser associated attributes with the
regex, not the field. We now no longer allow regex attributes as they
weren't used anywhere anyways. If we want to enable them again, we'll
need a new syntax.

Note: We still have a bug in parsing regexs with &size, see #286.