Fix `rt::Map::get`.
This patch removes the `default_` argument which could be used to
provide a value to return in case the requested key was not present in
the map.

We now instead return an optional reference to the contained value.
Users are expected to implement handling for the empty case themself.

We keep the HILTI and Spicy API consistent by maintaining the existing
signature through changed codegen.