Make autogen-docs script more robust.
We now first stage the new files, and copy them in place only at the
end once everything succeeded. We also use "rsync" to mirror the files
over only if changed, so that we don't mess with timestamps of the
existing files.

W don't report an error exit code if spicy-doc fails, so that one can
build the docs even without compiling.

I'm assuming that rsync is commonly installed on systems. There's a
check in the script to make sure we have it.