Fix comment parsing in EVT files.
The fix is quite involved because the parser is really just a hack
right now, and we need to add state machine to track whether we are
inside a comment. We need to switch to a flex/bison parser here
eventually.

We use new comments inside tftp.evt as our test, that used to be not
working.

Closes #312.