Rework dependencies so that Zeek plugin does no longer depend on compiler libraries when compiled with --disable-jit.
That means that the Zeek plugin will depend only on system libraries,
and can be feely copied around to other systems using the sam.
platform. No Spicy installation necessary as long as no JIT is needed.

This is a first commit just makes things work, with plenty of #ifdef.
Will refactor in a subsequent commit.

#299