Clean up the expression mode in the Spicy Bison parser.
The "expression mode" wasn't really doing what it sounded like: it was
almost always active, even if not inside an expression. Now we only
activate it when parsing an expression, so that we could take
decisions based on wheterh we're inside or outside a regexp. This
isn't currently used, but it's cleaner and will be potentially useful
later.

This comes with a couple of grammar changes to avoid ambiguitues:

    - "new X" no longer allowes full expressions for X, just values or
      identifiers.

    - Switch cases that use look-ahead now also need to be introduced
      with an arrow, e.g.:

           switch {
               -> a: A;
               -> b: B;
               -> c: C;
           };