Restrict parsing of attributes and properties to avoid ambiguities.
For properties, we limit the scanner to look for them only outside of
expressions.

For attributes, it's more tricky because our parser can't distinguiush
if an expression "x &y" means "x & y" or "x" followed by an attribute
"&y". In particular, field definitions are problematic in this regard,
e.g. "... &length=5 &chunked": What if "chunked" was a unit parameter?
We now address this by parsing "&y" as an attribute only if we know
that word as such.

Closes #350.