Allow access to unit parameters inside default expressions for unit variables.
This is slightly more complex than one would hope because we didn't
have access to the "__self" expression that accessing unit parameters
requires at the place where variables get initialized. We're solving
that by adding new method "__self()" to structs that returns what we
need. That then in turn allows to simplify some of the existing
"__self" setups a little bit.

Closes #355.