Type check assignments.
This also turned up a problem with propagating constness information:
when an operator's constness depended on a type deeper inside the AST
that could still change as more nodes got resolved, the operator's
constness could end up stale. The fix is to always calculate an
operator's result type dynamically from the current state of its
operands; and not cache it the first time, as we used to do. (This
change actually makes sense more broadly, wouldn't be surprised if it
could have triggered more issues elsewhere as well).

Closes #344. Closes #329.