Flipped implementation of `hilti::rt::Interval`-bool conversions.
The documentation and implementation of `hilti::rt::Interval`'s
conversion to bool were flipped. While the documentation stated the
"natural" semantics of non-empty intervals being truthy, the
implementation actually treated them as falsey and vice versa.

This patch brings the implementation in line with the documented
behavior. We also make the conversion `explicit` so it does e.g., not
participate in integer conversions.