Remove unintended conversion of `rt::Result` to `bool`.
`rt::Result` did provide implicit conversions to `bool`, but e.g., no
operators for `==` or `!=` so that Due to the language rules any (not)
equality check of `rt::Result` was performed by first converting both
operands to `bool` and then comparing them. Since `rt::Result`'s
bool conversion is just intended to check whether it has a value (a
typical use in conditionals), equality comparisions also just checked
whether both the operands had the same value state (both value, or both
error state). This is not what is intended.

This patch makes the conversion of `rt::Result` to `bool` `explicit`;
with that we can still use `rt::Result`s in conditionals, but disallow
unintended conversions. We also fix the code so that it now

- explicitly converts to bool where needed
- actually stringifies values in `rt::Result` instead of its value/error
  state
- tests for correct values