Use template specializations for hilti-rt tuple adaptors.
The way `operator<<` was declared made it impossible to find via ADL and
in general hard to work with. By using a template specialization in
`std` we can use more natural ADL to find the function. We move the
other functions in this file from SFINAE to ADL as well for consistency.