Switch `real::unpack` error reporting to use `rt::Result`.
While this function was already returning `Result`s and reporting
errors via `Error`s for some precondition violations, it was still
throwing exceptions for others. With this patch we use exclusively
`Result` to report errors up.

For unpacking of double precision numbers we also add a size check which
was already present for when unpacking single precision numbers.