Enforce some kind of include order with clang-format.
Previously our clang-format config did not include any Spicy-specific
include orderings, and we allowed authors to provide their own sections
which we then sorted alphabetically. This approach has lead to
inconsistencies in the code, e.g., no enforcement that header
corresponding to an implementation file was included first, or uncaught
incorrect groupings.

This patch updates the clang-format config to take a more aggressive
approach. We now predefine a number of Spicy-specific categories into
which we group files automatically. We also enforce a certain ordering
of includes from different categories. Since we regroup files we can now
automatically include corresponding headers first in implementation
files.

Since our include lists are often incomplete and depend on the incorrect
order this patch does not compile on its own. We will add missing in a
follow-up patch.