Fix memory issue when accessing map content.
I had cheated a bit and that was biting now: The code modelled map
elements as tuples, whereas they are really pairs (which we don't have
in HILTI). That happened to worked on macOS, but (not suprisingly) not
on Linux. The Map type-info class now does the right thing, although
it means duplicating quite a bit of code from the
IterableType---didn't see a nice way to unify that. On the upside, the
map type-info API is now nicer and more natural.