Fix build with static libraries.
The binaries weren't pulling in the Spicy plugin because the linker
didn't see any of its symbols referenced and would just strip it all
out, including the global constructor. Even when I forced my way
around that, the same problem would show up in other ways: other
constructors now missing, and runtime functions not being included
because a binary would seemingly not make use of them.

I now fixed that in a similar way as we already do for the runtime
libraries: using object libraries instead of static libraries. That
seems the way to go with CMake. So if the build is configured for
static libraries, we now link our binaries against separately defined
object versions of libhilti and libspicy. That makes sure everything
gets included. We still build the actual static libraries as well for
use by external applications.

This commits changes the CI build for clang9_zeek30_ubuntu_release to
use static libraries.

Just for the record, static linking is really slow, in particular on
Linux.

Note: Different from what I thought at first, I don't think we
actually need to do anything about spicy-config output because that's
not refering to libhilti/spicy at all, just to the runtime libraries
(which are already static).

Closes #400.