Make &cxxname usable for user code.
The code generator didn't produce a prototype for functions with
"&cxxname" yet. So far that worked because the attribute was only used
with internal library functions that already had prototypes in runtime
headers. For custom functions, however, there's no way to provide such
a prototype, so we now generate one automatically.

There's a catch though: For some of the library functions, HILTI
cannot generate a prototype because there's no direct mapping for some
of their arguments' types; we leverage some internall C++ knowledge
there. Therefore, there's now a new attribute that can be combined
with "&cxxname": "&have_prototype" disables the generation of a
prototype, assuming it's already available somewhere else. All the
library functions use that now.

Finally, we change the semantics of the ID provided with &cxxname to
to no longer requires a leading "::". They are now always assumed to
be absolute IDs with paths starting a global scope.

Documentation of this forthcoming in chapter on integration Spicy in
custom host applications.