Remove conversion of `rt::Time` to `bool`.
It was previously possible to check whether a `rt::Time` value was
"unset" by converting it to a `bool`. "Unset" here means a `Time` value
set to the `null` Time value (the zero point of the epoch).

The operator was implemented incorrectly (returning the inverse of the
expected result). We remove it here since the intended "unset" `Time`
value is in fact a valid `Time` value. To check whether a `Time` value is
different from the zero point of the epoch users should instead directly
compare `Time` values against null values if it fits their intended
semantics.