Allow limiting the %port of a unit to a specific direction.
We now allow to add attributes `&originator`/`&responder` to a `%port`
unit property to limit it to the corresponding direction of a session.
When selecting a parser through `--parser <port>` in spicy-driver, one
can now add `%orig` or `%resp` to the port to indicate the desired
directionality. If one leaves that off, only parsers are selected that
have no directionlity defined (i.e., just like it used to be).

The main task for the implementation here is to channel the direction
attributes through into the runtime library. For that, we follow
pretty closely what we're already doing for MIME types.

The Zeek script to create Spicy batches now adds the direction
information to the sessions' ports as well.