Embed version information into HLTO files.
We now add a static JSON string to all HLTO files that encodes
information about the HILTI version/config they were built with. We
use this for detecting (1) mismatches in the HITLI version used for
building vs running and (2) running an optimized HLTO file with a
debug runtime. In both cases, the runtime will print a warning to
stderr, but still proceed processing the library. We might want turn
at least (1) into a fatal error at some point, let's see.