Add `hilti::rt::Library::symbol`.
This patch adds a new member function `Library::symbol` which can be
used to obtained a symbol from a specific library.

While currently `Library` users just call `::dlsym` globally to obtain a
symbol, this becomes ambigous if multiple libraries in the process image
provide similarly named symbols. Instead they could hold handles to
all interesting libraries and obtain symbols with the added function.
With that we could also `::dclose` libraries in the `Library`
destructor; we'll leave that change for a later point.