Fix include order of quote includes.
clang-format already figures out the main header correctly and in the
current config included it first, and the regex we had in place for it
also collected any other quote includes.

This patch introduces a dedicated include category for quote includes
which is placed at the end of the include list.