Make test of `resource_usage` atomic.
This test was previously using whatever environment the previous tests
left behind to perform tests. This limited both how much we could narrow
down the range of expected values and also made it impossible to run
this test in isolation.

With this patch this now sets up a controlled environment so it can be
run in isolation and more constrain the statistics under test.