Always link against thread library.
The performance of some of our runtime code depends on whether the
executable is linked against pthread or not (because libstdc++ then
switches to thread-safe versions for some stuff). To ensure our
performance is consistent, we now always link against the system's
thread library, even if that reduces overall performance.