Move resolving of exceptions' backtrace symbols to when we need them.
We used to fully resolve symbols of the current stack backtrace
whenever we instantiated an exception, independent of whether we would
ever print the symbols. Turns out that that can be very expensive, in
particular when the code sits inside a shared library it caused a
substantial runtime overhead. We now just capture the raw addresses at
exception instantiation, and resolve them only when we're going to
print the backtrace (which is almost never).