Reorganize source code to allow for building the runtime library standalone.
This splits out what used to be the "hilti/" tree, into two separate
blocks "hilti/toolchain/" and "hilti/runtime/". The latter contains
what used to be in hilti/{include,src}/rt". Along the same lines, this
likewise splits "spicy/" into "spocy/toolchain/" and "spicy/runtime/".

Doing so then allows us to update the CMake config so that the
"{hilti,spicy}/runtime" subdirectories can be built independent of the
toolchain. "configure" gains a corresponding option to turn off
building the toolchain.

This commits also extends hilti-config and spicy-config with options
to print the runtime's C++ include directories.