Turn all 3rdparty code into proper git submodules.
This adds a couple of work-arounds for issues that we had previoulsy
directly patched inside the 3rd party code.

This upgrades all 3rdparty code to the most recent versions, and moves
LICENSE.3rdparty into 3rdparty/.

For the base64 code, the version upgrade allows (and requires) removal
of a work-around for a previous bug.

justrx now lives it its own repository (on my personal account for
now, not quite sure yet what to do with it).