Add CMake module that enables Zeek plugins to build Spicy parsers.
This will be installed along with Spicy's Zeek plugin so that other
Zeek plugins can pull it in.

This commit comes with a few small tweaks to the plugin to make this
all work. One of those changes is a renaming of the plugin from
"Zeek::Spicy" to "_Zeek::Spicy". That's a hack to make sure Zeek loads
our plugin early on so that the runtime is available once any
analyzers get loaded (Zeek loads plugins in alphabetical order).