Add support for packet analyzers to Zeek plugin.
Zeek master got new support for adding low-level "packet analyzers"
through plugins. This extends the Spicy plugin to support that. We now
allow a new section "packet analyzer" in *.evt files that defines the
unit to use for parsing a packet:

    packet analyzer spicy::RawLayer:
        parse with RawLayer::Packet;

This is all needed to make the new analyzer available to Zeek. Usually
you will need to add two more pieces to make it useful:

1. If the analyzer just strips away a header to then forward the
   remaining data to another analyzer, you need to call
   ``zeek::forward(protocol)`` at some point during the unit's
   parsing. `protocol` is the identifier of the next protocol (per
   Zeek's dispatcher tables).

2. You need to configure the Zeek-side dispatcher tables, similar to
   any other packet analyzers; see the documentation.

This needs Zeek master currently.

Full example in tests/zeek/evt/packet-analyzer.Zeek.

No CI integration so far as we aren't testing Zeek master (4.0-to-be).

One additional backwards-incompatible change: we no longer normalize the
name of file/protocol, meaning that an "enable" in a signature needs to use
the original colon-separated name now, and no longer replaces the colons
with an underscore. Docs have been updated.

Closes #497.