Do not create copies of shared libraries.
We were initially copying shared libraries to a store in the system temp
directory so that we could treat JIT'ed libraries and existing libraries
compiled ahead of time similarly.

In this patch we remove the notion of that store and instead directly
works with the libraries. We remove code which deletes library files
when their handles go out of scope; instead only JIT'ed libraries get
GC'ed when they go out of scope while this is not done for existing
libraries.

Closes #521.