Remove --include-linker option from driver.
This option used to print the linker glue code along with the main,
generated C++ code. The problem is that the linker code was produced
as if it were a standalone C++ unit. However, by concatening it with
the main code, that didn't hold and we'd run into problems like
duplicaste type definitions. This wasn't salvagable easily, and
conceptually it looks kind of wrong anyways, so removing it.

The better way to do all this is using "-l" to generate *just* the
linker code, then compile the two together as two compilation units.
If one insists, one can also still "cat" the two and will be back to
what "-cK" used to produce, with the same issues obviously.

Closes #500.