Remove packaging tasks for static CI build.
Static builds tend to produce very large binaries which cannot reliably
be compressed to be inside Cirrus CI's 1GB artifact size limit.

This patch removes the packaging step for the
clang9_zeek_lts_ubuntu_release_static build task. The package artifacts
of this task are currently not referenced in the documentation.