Optionally precompile libhilti.h.
This patch adds precompilation of libhilti.h for developer setups. This
header is included in all files during JIT, and precompiling it can for
individual files reduce JIT time by up to 30%. This behavior can be
disabled by setting the CMake option `HILTI_DEV_PRECOMPILE_HEADERS=OFF`.

Specifically, we add a cmake target which creates debug and release
versions of a precompiled libhilti.h. We modify HILTI's built-in
CXXFLAGS to check for the presence of the needed precompiled header in
the expected location. We expect the header either in a specific
location in the build directory, or optionally in a version directory in
`$HOME/.cache/spicy`.

In this patch we enable use of the precompiled header only during JIT,
but e.g., not when building Spicy targets. This allows us to profit from
faster JIT while still making efficient use of e.g., ccache.
Additionally, this approach allows us to punt on fixing a number of
issues in header files which get exposed during JIT (JIT artifacts are
not affected at the moment).

The original goal was to precompile lihilti.h before installation which
is not reached. In general, it seems that precompiled headers are not
intended to be moved after creation. This poses a challenge when the
header gets installed: while we could in principle generate the
precompiled header at the final location during install time, this seems
not reliably possible when generating packages which can be expanded in
arbitrary locations.

Additionally, we also install a script for users to take advantage of
precompiled headers once Spicy has been installed.