Extend while-production to support termination of loop by look ahead.
So far a while-production required an explicit condition to terminate
the loop (which it turns out isn't actually used anywhere right now).
This adds a 2nd version that doesn't need an expression and instead uses
look-ahead parsing to decide when to terminate the loop: If the next
input token indicates that we can match another instance of the loop's
body, we go there; if the next token indicates that we can match
whatever comes after the loop, we terminate; and if both or none of this
is the case, that's an error. Internally, we create a little wrapper
grammar to implement this, so that global look-ahead computation works
as normal across such a while-loop.

This logic is essentially all the same as what the grammar builder is
currently doing already for parsing a vector through look-ahead---and
we're going to replace that existing logic with this new loop in a
subsequent commit.