Switch Fiber runtime to fiber.
This patch moves the Fiber runtime from libtask to the fiber library.
While the changes here are just an internal refactoring which do not
enable new functionality, having the execution contexts of code path
more explicit should make the code easier to understand; we might also
be able to implement a shared stack on top of the fiber library so we
can be less pessimistic about stack memory requirements in the future.