[WIP] Move fiber runtime to libaco.
This patch is a collection of smaller changes which move the running of
fibers from libtask to libaco. We still need to break this patch into
smaller pieces in a cleanup.

While on the surface this should be a simple refactoring changing uses
of `_setjmp` and `_longjmp` to `aco_yield` and `aco_resume` we require
another change in Fiber::run semantics; while it is possible to jump
between arbitrary coroutines with raw jumps, with libaco one can only
yield from a coroutine to the global main coroutine, and only resume
from that main coroutine. We add some preliminary notion of fiber calls
stacks to the HILTI runtime to accomplish that. Currrently that stack is
maintained from `Fiber::run`. Jumps between coroutines are composed of

- putting the next coroutine on the stack,
- yielding to main,
- resuming a fiber which first executes all pending coroutines and only
  then resumes the current coroutine.

This approach changes fiber semantics and is now e.g., possible to
resume a running fiber (a fiber it called while running might have
yielded itself, leaving only the called coroutine in yielded state).

More, it looks like we will be required to add another fiber management
layer around that the current runtime which ensures that fibers finish
running (e.g., resuming them until the reach a terminal state). Since we
do not have these changes a number of tests still fail.

This patch also contains changes making sure that the libaco runtime
environment is correctly set up when reinitializing the HILTI runtime
lib, and any lingering fiber stack state is cleaned up.