Mark two AST node classes as singletons.
This means we don't need to deep-copy them, but can just share
references across instances. This is generally possible for nodes that
will never have any children (because that means they won't change at
all).

(Note to future self: This change is useful only for classes that are
directly wrapped into a `Node`, not for classes that are type-erased
through further intermediary type-erased parents (e.g., types,
expression). In other words: only for classes that have the trait
``isNode``. The reason is that only `Node` does deep-copies anyways, the
other type-erased parents always do shallow copies.)