Fix argument mismatches between Spicy generated C++ prototypes and Zeek plugin runtime code.
Some signatures in the Zeek plugin runtime code were simply wrong. This
has been in there like this for a while already. It's independent of the
rest of this branch, but for some reason became an issue only now. Not
sure why we haven't seen this trigger earlier already.

Also turning a bunch of std::string_view into std::string for
consistency.