Add capture group extraction to regexp matcher state.
This adds a new method to `MatchState` for retrieving a vector of
captured groups after a successful match. The change is a bit more
involved than one would hope because it revealed a couple of existing
deficiencies:

1. `MatchState` was previously limited to using JRX's minimal matcher,
   which doesn't support capturing. When changing that it turned out
   that we were inconsistent in how to get to the full match across the
   two types of matchers. This is being fixed here, also requiring the
   JRX change pulled in in the previous commit.

2. Using `&nosub` used to imply anchoring an expression as well, which
   led to further inconsistencies, also with respect to other regexp
   methods. We now split that up into two flags. (This will require
   changes to a number of HILTI and Spicy tests, which now need to add
   `&anchor`. Those will come in follow up commits.)

This might be hard to review because it's quite low-level. Probably
best to focus on the new unit tests that aim to exercise the various
combinations of flags and input mode.