Make new captures() method available on regexp matching state in HILTI.
This also introduces a new keyword expression to refer to the array of
matching groups. We don't actually use that HILTI-side, but Spicy will
(it's similar to the other keyword types in this regard).

One special thing here: we internally use "$@" as the syntax to access
that array. However, we won't make that special identifier available to
users (not even in Spicy). Instead, we'll introduce $1/$2/... later.