When passing control into a new fiber callback, copy all arguments into the heap.
Code now looks like this example:

    extern auto hlt::Foo::test(const std::string& x) -> hilti::rt::Resumable {
        auto args = std::make_tuple(::hilti::rt::resumable::detail::copyArg(x));
        auto args_on_heap = std::make_shared<decltype(args)>(std::move(args));
        auto cb = [args_on_heap](hilti::rt::resumable::Handle* r) -> std::any {
            return __hlt::Foo::test(std::get<0>(*args_on_heap));
        };

        auto r = std::make_unique<hilti::rt::Resumable>(std::move(cb));
        r->run();
        return std::move(*r);
    }

We special-vase `ValueReference` because that's already on the heap, and
we can just do a shallow copy. We also change the parse3() function to
parse the `ParsedUnit` through a `ValueReference` to get the same
shallow copying behaviour (which we need because the instance is being
modified).

Note: We could optimize this further. The arguments don't necessarily
need to go on the heap, they just need to remain available when we
switch out the stack.