Add shared stack and refactor fiber code.
This adds the trampoline doing the stack switching, and generally
refactors state mgmt & control flow quite a bit further. We now keep all
fiber-related non-local state inside the execution context. That state
is managed fully from fiber.cc. This includes the fiber cache, which has
moved into the context as well.

We use the shared stack by default now, unless we're compiling with
ASAN: ASAN doesn't like the memcpys to save/restore the stack and there
doesn't seem to be a way to suppress it flagging. So in ASAN mode, we
use fibers with individual stacks instead, but still go through the
switcher trampoline to keep the code path as similar as possible.