Be explicit on which variant we construct in `Result`.
We were relying on the fact that we constructed the values in a
`Result` without conversions. This meant that the variant constructors
didn't need to specify which variant was being constructed.

This patch makes it explicit in `Result` constructors which variant we
construct. This helps us avoid running down into potentially ambiguous
conversion checks as seen e.g., on Debian setups.