Overhauling regexp matching semantics.
We were inconsistent in how regular expressions (1) handled being
anchored vs not, and (2) allowed to retrieve the matching data. The main
underlying problem was that we determined at regexp compile time which
of the JRX matchers to use: the standard matcher (which allows to
retrieve the matching part as well as all capture groups), or the
minimal matcher which matches more efficiently but allows to retrieve
only the final position of a match (and no capture groups). Likewise, we
so far had to decide at compile time whether to consider an expression
anchored not not. These decisions upfront then meant that the actual
matching methods had to work with what they were handed, leading to
unintuitive restrictions and/or lack of efficiency.

This commits changes this as follows:

- We always compile regexps for the standard matcher (which the minimal
  matcher can use as well).

- We always compile regexps as anchored, no implicit `.*` at compile
  time anymore.

- We let the matching methods decide which type of matcher to choose
  (this required an update to justrx to extend its API accordingly; this
  commit includes the corresponding submodule update).

- We rename the regexp methods for more consistency: `match` now means
  "matching is anchored at the beginning of the data", and `find` means
  we find matches anywhere in the data. This leads to the following
  renames:

    * <regexp>.find -> <regexp>.match
    * <regexp>.find_groups -> <regexp>match_groups
    * <regexp>.find_span -> <regexp>.find

  `<bytes>.match` adapts to the new semantics as well.

  Same renaming happens at the C++ level.

- We remove the `&anchor` attribute because that's the default now.

- `&no_sub` remains available to override the default choice of the
  matcher type: it'll use the minimal matcher and disable capture group
  support.

- The one downside of all this is that the new `find` is a bit
  inefficient: because all our regexps are compiled anchored, it needs
  to try matching at all possible begin positions, meaning quadratic
  runtime. Doesn't seem like a big issue right now, but we could improve
  this in the future by compiling two DFAs for each regexp: anchored and
  non-anchored, and then using the standard matcher instead of the
  minimal matcher with `find`.